Fixed issue with responses not being intercepted
Social experiment BeliefDisclosure responses were not being
correctly intercepted. This was related to the fact that
the source address of survey queries was not being set
to the SocialExperiment object itself.

Limited tests (not true unit tests, need to integrate JUnit tests)
were also added for convenient debugging of SocialExperiment.